+++ CHANGELOG.md
@@ -12,6 +12,7 @@ with respect to its command line interface and HTTP interface.
   pre-slug data instead of a random UUID.
 
 ### Major
+- Remove Annotation type from the sous package.
 - Singularity RequestIDs are shortened to no longer include FQDN or
   organization of Git repo URL.
 - DeploymentID receives the Digest() function.